Task 0 and 1 corrected unsined int to int

+++ singly_linked_lists/0-print_list.c
@@ -4,19 +4,19 @@
 #include <string.h>
 /**
  * print_list - print element of a list
- * @h: List linked
- * Return: size of list
+ * @h: pointer to the list
+ * Return: number of nodes
  */
 size_t print_list(const list_t *h)
 {
-	int len = 0;
+	size_t len = 0;
 
 	while (h != NULL)
 	{
 		if (h->str == NULL)
 			printf("[Ø] (nil)\n");
 		else
-			printf("[%d] %s\n", h->len, h->str);
+			printf("[%u] %s\n", h->len, h->str);
 	h = h->next;
 	len++;
 	}
